% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToR.R
\name{bamToR}
\alias{bamToR}
\title{bamToR}
\usage{
bamToR(filename)
}
\arguments{
\item{filename}{string contatinig name of file with bam data. Remember that you need to set path to folder with selected file.}
}
\description{
Converts .bam file into a R list
}
\examples{
\dontrun{
filename <- "P13.2016-11-17T16_40_36.1234062861011.fc1.ch1.sorted.bam"
bamData <- bamToR(filename)
}
}
