% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsPerGene.R
\name{countsPerGene}
\alias{countsPerGene}
\title{Counts for gene}
\usage{
countsPerGene(datawt = list(), databrm = list(), datac = list(), gene,
  geneData = directRNAExplorer::TAIR10_genes, genePart = "gene",
  type = "unique")
}
\arguments{
\item{datawt}{a list of bamDataFrames received after use the \code{bamToR} function containing information from the wt individual}

\item{databrm}{a list of bamDataFrames received after use the \code{bamToR} function containing information from the brm individual}

\item{datac}{a list of bamDataFrames received after use the \code{bamToR} function containing information from the c individual}

\item{gene}{a character to indicate for which gene we compute the counts}

\item{geneData}{a data.frame containing informations about genes}

\item{genePart}{a character to indicate for what gene part we want to compute the counts, by default it is the "gene"}

\item{type}{a character to indicate whether we count unique counts or all of them}
}
\description{
Function \code{countsPerGene} create a list with three data frames containing counts for gene or choosen gene part.
}
