% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgerTest.R
\name{edgerTest}
\alias{edgerTest}
\title{EdgeR based test for differential expression analysis}
\usage{
edgerTest(countsData, condition, type = "qlf", ...)
}
\arguments{
\item{countsData}{Data frame containing counts for each}

\item{condition}{Vector of conditions. In our data the type of mutation}

\item{type}{Type of test "lrt" for likelihood ratio tests, "qlf" for quasi-likelihood F-tests.}

\item{...}{Other arguments.}
}
\description{
Function \code{test_edger} computes the statistics for quasi-likelihood F-tests or likelihood ratio tests.
}
