% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksDistributionTest.R
\name{ksDistributionTest}
\alias{ksDistributionTest}
\title{ksDistributionTest}
\usage{
ksDistributionTest(data1, data2, gene,
  geneData = SequencingExplainer::TAIR10_genes, strand = "pos",
  genePart = "gene")
}
\arguments{
\item{data1}{First data set which contains informations about selected gene.}

\item{data2}{Second data set which contains informations about selected gene.}

\item{gene}{Gene name}

\item{geneData}{Data frame with positions of all genes and their names, by default we use a \code{TAIR10_genes_tidy}}

\item{strand}{On which strand we want to compute test}

\item{genePart}{The part of gene we want to test. By default we conduct test for whole gene.}
}
\description{
Kolmogorov-Smirnov test for gene from two datasets
}
