% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinomTest.R
\name{nbinomTest}
\alias{nbinomTest}
\title{Negative Binomial Test}
\usage{
nbinomTest(countsData, condition, ...)
}
\arguments{
\item{countsData}{Data frame containing counts for each}

\item{condition}{Vector of conditions. In our data the type of mutation}

\item{...}{Other arguments.}
}
\description{
Negative Binomial Test
}
