% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{plotCoverage}
\usage{
plotCoverage(bamDataFrame, chromosome, start, stop, ...)
}
\arguments{
\item{bamDataFrame}{Data frame converted to R using \code{bamToR()} function}

\item{chromosome}{Number of chosen chromosome}

\item{start}{First position in gene/exon/three prime utr etc.}

\item{stop}{Last position in gene/exon/three prime utr etc.}

\item{...}{Optional arguments}
}
\description{
Coverage plot for chosen part of RNA sequence
}
