% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneComparisonDistribution.R
\name{plotGeneComparisonDistribution}
\alias{plotGeneComparisonDistribution}
\title{plotGeneComparisonDistribution}
\usage{
plotGeneComparisonDistribution(gene, data1, data2,
  geneData = SequencingExplainer::TAIR10_genes, range = 0,
  genePart = "gene", adjust = 1, ...)
}
\arguments{
\item{gene}{Character with gene name.}

\item{data1}{Data frame converted to R using \code{bamToR()} function, corresponding to the first sample.}

\item{data2}{Data frame converted to R using \code{bamToR()} function, corresponding to the second sample.}

\item{geneData}{Data frame with positions of all genes and their names, by default we use a \code{TAIR10_genes}.}

\item{range}{How many nucleotide before \code{start} and after \code{stop} we include to genes.}

\item{genePart}{The part of gene we want to visualize.}

\item{adjust}{Adjust parameter for the density kernel}

\item{...}{Optional arguments.}
}
\description{
Plot of distribution of two samples on selected gene
}
