% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoints.R
\name{plotPoints}
\alias{plotPoints}
\title{plotPoints}
\usage{
plotPoints(gene, data1, data2, data3, geneData = directRNAExplorer::Araport11,
  range_st = 0, range_end = 0)
}
\arguments{
\item{gene}{Character with gene name.}

\item{data1}{Data frame converted to R using \code{bamToR()} function.}

\item{data2}{Data frame converted to R using \code{bamToR()} function.}

\item{data3}{Data frame converted to R using \code{bamToR()} function.}

\item{geneData}{Data frame with positions of all genes and their names, by default we use a \code{Araport11}.}

\item{range_st}{How many nucleotides before \code{start} include to genes.}

\item{range_end}{How many nucleotides after \code{stop} we include to genes.}
}
\description{
plotPoints
}
