% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalTable.R
\name{pvalTable}
\alias{pvalTable}
\title{pvalTable}
\usage{
pvalTable(data1, data2, chromosome,
  geneData = directRNAExplorer::TAIR10_genes, type = "KS",
  genePart = "gene", strand = "+", ...)
}
\arguments{
\item{data1}{First data set which contains informations about selected gene.}

\item{data2}{Second data set which contains informations about selected gene.}

\item{chromosome}{Chromosome name.}

\item{geneData}{Data frame with positions of all genes and their names, by default we use a \code{TAIR10_genes}.}

\item{type}{Type of chosen test}

\item{genePart}{The part of gene we want to visualize.}

\item{strand}{On which strand we want to compute test.}

\item{...}{Optional arguments}
}
\description{
Create data.frame with pvalues and value of statistic for chosen statistical test
}
