% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\name{tTest}
\alias{tTest}
\title{T-student test}
\usage{
tTest(data, condition, ...)
}
\arguments{
\item{data}{Data set containing counts for each gene, columns corresponds to genes, rows to samples}

\item{condition}{A two - levels vector of condition values}

\item{...}{Other arguments}
}
\description{
Function \code{tTest} computes statistics for genes with division into two groups of probes.
}
