% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchToGene.R
\name{matchToGene}
\alias{matchToGene}
\title{matchToGene}
\usage{
matchToGene(positions, start, stop, geneName, range = 0)
}
\arguments{
\item{positions}{Positions from sequencing.}

\item{start}{Gene start index.}

\item{stop}{Gene stop index.}

\item{geneName}{Gene name.}

\item{range}{How many nucleotide before \code{start} and after \code{stop} we include to genes.}
}
\description{
Counts how many sequencing points were at known genes
}
\examples{
library(SequencingExplainer)
data <- dataChromosome1
dic <- TAIR10_genes
matchToGene(positions=unique(data$pos), start=dic$V4, stop=dic$V5, geneName=dic$id)

}
