% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchToGene.R
\name{matchToGene}
\alias{matchToGene}
\title{matchToGene}
\usage{
matchToGene(positions, start, stop, geneName, range = 0)
}
\arguments{
\item{positions}{positions from sequencing}

\item{start}{gene start index}

\item{stop}{gene stop index}

\item{geneName}{gene name}

\item{range}{how many nucleotide before \code{start} and after \code{stop} we include to genes}
}
\description{
Counts how many sequencing points were at known genes
}
