% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{plotCoverage}
\usage{
plotCoverage(bamDataFrame, chromosome, ...)
}
\arguments{
\item{bamDataFrame}{data frame converted to R using \code{bamToR()} function}

\item{chromosome}{number of chosen chromosome}

\item{...}{optional arguments}
}
\description{
Coverage plot for chosen part of RNA sequence
}
