% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneDistribution.R
\name{plotGeneDistribution}
\alias{plotGeneDistribution}
\title{plotGeneDistribution}
\usage{
plotGeneDistribution(gene, bamDataFrame,
  geneData = SequencingExplainer::TAIR10_genes, range = 0,
  genePart = "gene", ...)
}
\arguments{
\item{gene}{Character with gene name}

\item{bamDataFrame}{Data frame converted to R using \code{bamToR()} function}

\item{geneData}{Data frame with positions of all genes and their names, by default we use a \code{TAIR10_genes_tidy}}

\item{range}{How many nucleotide before \code{start} and after \code{stop} we include to genes}

\item{genePart}{The part of gene we want to visualize.}

\item{...}{Optional arguments}
}
\description{
Plot of distribution on selected gene
}
